pyproject.toml
[project]
name = "inforecon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiodns>=3.5.0",
    "aiohttp>=3.13.1",
    "crawl4ai>=0.7.6",
    "dnspython>=2.8.0",
    "fastapi>=0.119.1",
    "geoip2>=5.1.0",
    "playwright>=1.55.0",
    "pydantic-settings>=2.11.0",
    "python-whois>=0.9.6",
    "redis>=7.0.1",
    "uvicorn[standard]>=0.38.0",
]
